\name{methods.escalc}
\alias{methods.escalc}
\alias{[.escalc}
\alias{cbind.escalc}
\alias{rbind.escalc}
\title{Methods for 'escalc' Objects}
\description{
   Methods for objects of class \code{"escalc"}.
}
\usage{
\method{[}{escalc}(x, i, \dots)
\method{cbind}{escalc}(\dots, deparse.level=1)
\method{rbind}{escalc}(\dots, deparse.level=1)
}
\arguments{
   \item{x}{an object of class \code{"escalc"}.}
   \item{\dots}{other arguments.}
}
\author{
   Wolfgang Viechtbauer \email{wvb@metafor-project.org} \url{http://www.metafor-project.org/}
}
\references{
   Viechtbauer, W. (2010). Conducting meta-analyses in R with the metafor package. \emph{Journal of Statistical Software}, \bold{36}(3), 1--48. \url{http://www.jstatsoft.org/v36/i03/}.
}
\keyword{internal}
