\name{print.list.rma}
\alias{print.list.rma}
\title{Print Method for list.rma Objects}
\description{
   Print method for objects of class \code{"list.rma"}.
}
\usage{
\method{print}{list.rma}(x, digits=x$digits, \dots)
}
\arguments{
	\item{x}{an object of class \code{"list.rma"}.}
	\item{digits}{an integer specifying the number of decimal places to which the printed results should be rounded (the default is to take the value from the object).}
	\item{\dots}{other arguments.}
}
\value{
	See the documentation of the function that creates the \code{"list.rma"} object for details on what is printed.
}
\author{Wolfgang Viechtbauer; \email{wvb@www.wvbauer.com}; \url{http://www.wvbauer.com/}}
\keyword{print}
