\name{print.rma.list}
\alias{print.rma.list}
\title{Print Method for rma.list Objects}
\description{
   Print method for objects of class \code{"rma.list"}.
}
\usage{
\method{print}{rma.list}(x, digits=x$digits, ...)
}
\arguments{
	\item{x}{an object of class \code{"rma.list"}.}
	\item{digits}{an integer specifying the number of decimal places to which the printed results should be rounded (the default is to take the value from the object).}
	\item{...}{other arguments.}
}
\value{
	See the documentation of the function that creates the \code{"rma.list"} object for details on what is printed.
}
\author{Wolfgang Viechtbauer; \email{wvb@www.wvbauer.com}; \url{http://www.wvbauer.com/}}
\keyword{print}
