\name{print.rma.uni.infl}
\alias{print.rma.uni.infl}
\title{Print Method for rma.uni.infl Objects}
\description{
   Print method for objects of class \code{"rma.uni.infl"}.
}
\usage{
\method{print}{rma.uni.infl}(x, digits=x$digits, ...)
}
\arguments{
	\item{x}{an object of class \code{"rma.uni.infl"}.}
	\item{digits}{an integer specifying the number of decimal places to which the printed results should be rounded (the default is to take the value from the object).}
	\item{...}{other arguments.}
}
\value{
	See the documentation of the \code{\link{influence.rma.uni}} function for details on what is printed.
}
\author{Wolfgang Viechtbauer; \email{wvb@www.wvbauer.com}; \url{http://www.wvbauer.com/}}
\seealso{
	\code{\link{influence.rma.uni}}
}
\keyword{print}
