\name{reporter}
\alias{reporter}
\alias{reporter.rma.uni}
\title{Dynamically Generated Analysis Reports for 'rma.uni' Objects}
\description{The function dynamically generates analysis reports for objects of class \code{"rma.uni"}.}
\usage{
reporter(x, \dots)

\method{reporter}{rma.uni}(x, dir, filename, format="pdf_document", open=TRUE,
         digits, forest, funnel, footnotes=FALSE, verbose=TRUE, \dots)
}
\arguments{
   \item{x}{an object of class \code{"rma.uni"}.}
   \item{dir}{optional character string specifying the directory for creating the report (by default, \code{\link{tempdir}} will be used).}
   \item{filename}{optional character string specifying the filename (without file extension) for the report (by default, the function sets a filename automatically).}
   \item{format}{output format for the report (either \code{pdf_document}, \code{html_document}, or \code{word_document}). Can be abbreviated.}
   \item{open}{logical indicating whether the report should be opened after it has been generated (default is \code{TRUE}).}
   \item{digits}{integer specifying the number of decimal places to which the printed results should be rounded (if unspecified, the default is to take the value from the object).}
   \item{forest}{optional character string with arguments to be added to the call to \code{\link{forest.rma}} for generating the forest plot.}
   \item{funnel}{optional character string with arguments to be added to the call to \code{\link{funnel.rma}} for generating the funnel plot.}
   \item{footnotes}{logical indicating whether additional explanatory footnotes should be added to the report (default is \code{FALSE}).}
   \item{verbose}{logical indicating whether information on the progress of generating the report should be provided (default is \code{TRUE}).}
   \item{\dots}{other arguments.}
}
\details{
   The function dynamically generates an analysis report based on the model object. The report includes information about the model that was fitted, the distribution of the observed outcomes, the estimate of the average outcome based on the fitted model, tests and statistics that are informative about potential (residual) heterogeneity in the outcomes, checks for outliers and/or influential studies, and tests for funnel plot asymmetry. A forest plot and a funnel plot are also provided.
}
\value{
   The function generates either a pdf, html, or docx file and returns the path to the generated document.
}
\note{
   Since the report is created based on an R markdown document that is generated by the function, the \code{rmarkdown} package (\url{https://cran.r-project.org/package=rmarkdown}) and \code{pandoc} (\url{http://pandoc.org}) must be installed.

   If \code{filename} is unspecified, the default is to use \code{report}, followed by an underscore (i.e., \code{_}) and the name of the object passed to the function. Both the R markdown file (with extension .rmd) and the actual report (with extension .pdf, .html, or .docx) are named accordingly. To generate the report, the model object is also saved to a file (with the same filename as above, but with extension .rdata). Also, files \code{references.bib} and \code{apa.csl} are copied to the same directory (these files are needed to generate the references in APA format).

   Since the report is put together based on predefined text blocks, the writing is not very elegant. Also, using personal pronouns (\sQuote{I} or \sQuote{we}) does not make sense for such a report, so a lot of passive voice is used.

   The generated report provides an illustration of how the results of the model can be reported, but is not a substitute for a careful examination of the results.
}
\author{
   Wolfgang Viechtbauer \email{wvb@metafor-project.org} \url{http://www.metafor-project.org/}
}
\references{
   Viechtbauer, W. (2010). Conducting meta-analyses in R with the metafor package. \emph{Journal of Statistical Software}, \bold{36}(3), 1--48. \url{http://www.jstatsoft.org/v36/i03/}.
}
\seealso{
   \code{\link{rma.uni}}
}
\examples{
### load BCG vaccine data
dat <- get(data(dat.bcg))

### calculate log risk ratios and corresponding sampling variances
dat <- escalc(measure="RR", ai=tpos, bi=tneg, ci=cpos, di=cneg, data=dat,
              slab=paste(dat$author, ", ", dat$year, sep=""))

### random-effects model
res <- rma(yi, vi, data=dat)

\dontrun{
### generate pdf report
reporter(res)}
}
\keyword{methods}
