\name{metafor.news}
\alias{metafor.news}
\title{Read News File of the Metafor Package}
\description{
   Read news file of the \pkg{\link{metafor-package}}.
}
\usage{
metafor.news()
}
\details{
   The function is just a wrapper for \code{news(package="metafor")} which parses and displays the \file{NEWS} file of the package.
}
\author{
   Wolfgang Viechtbauer \email{wvb@metafor-project.org} \url{http://www.metafor-project.org}
}
\references{
   Viechtbauer, W. (2010). Conducting meta-analyses in R with the metafor package. \emph{Journal of Statistical Software}, \bold{36}(3), 1--48. \url{https://www.jstatsoft.org/v036/i03}.
}
\examples{
\dontrun{
metafor.news()}
}
\keyword{utilities}
